Add comments to rpi_disp.c
The file rpi_disp.c holds stubs for accelerated 2D blit and fill
primitives that might be implemented in the future for the rpifb
driver.

I am aware of an advanced X driver for the RPi that uses the DMA
engines for this purpose. However, checking out the source
code of this driver suggests it uses busy-waiting for completion,
although able to start off DMA requests asynchronously. For stability,
it might be advantageous to start with an implementation that
performs the entire fill or blit request inside the kernel, with the
kernel sleeping until an interrupt occur signalling that the operation
has finished.