Optimize conditionals in X driver fill function
Probably due to branch-prediction characteristics of the ARM CPU, it is
sensitive to the way conditional statements are ordered. This commit
set the fill function to NULL if it is not implemented instead of having
a no-op functions returning 0. The fill function of the X-driver tries
to minimize the conditional penalties, resulting in improved performance
for small off-screen fills:

PixmapFillRect (5 x 5): Speed up 26%
PixmapFillRect (7 x 7): Speed up 30%
PixmapFillRect (10 x 10): Speed up 25%
PixmapFillRect (15 x 15): Speed up 8%
PixmapFillRect (22 x 22): Speed up 6%